[SORRY] i made a complete rewrite in a separate non-git directory and here it is in one giant commit..
It fixes a serious issue that happened if a workspace had a mark and no tiled windows. If a window was later set to be tiled on that workspace things got crazy. This is due to a bug in i3 itself, and the fix here is to not mark the workspace.

Another notable change in this version is that it uses zero (0) i3-msg open commands, now "ghosts".